Library revert, manual approach
Stencil library turned out to be too much of a hassle to get it working with the Electron+Typescript combo, so I removed it. Current approach is a (far too complex) manual class extension, but until (or if no) alternatives are found, it will have to stay that way.
The current push includes a sample component that is both stateful and controllable.